Reorg map loading classes
