+++ src/map-loader/GidTidTranslator.cpp
@@ -1,0 +1,76 @@
+/******************************************************************************
+
+    GPLv3 License
+    Copyright (c) 2023 Aria Janke
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+*****************************************************************************/
+
+#include "GidTidTranslator.hpp"
+#include "TileSet.hpp"
+
+namespace {
+
+using namespace cul::exceptions_abbr;
+using ConstTileSetPtr = GidTidTranslator::ConstTileSetPtr;
+
+} // end of <anonymous> namespace
+
+GidTidTranslator::GidTidTranslator
+    (const std::vector<TileSetPtr> & tilesets, const std::vector<int> & startgids)
+{
+    if (tilesets.size() != startgids.size()) {
+        throw RtError{"Bug in library, GidTidTranslator constructor expects "
+                      "both passed containers to be equal in size."};
+    }
+    m_gid_map.reserve(tilesets.size());
+    for (std::size_t i = 0; i != tilesets.size(); ++i) {
+        m_gid_map.emplace_back(startgids[i], tilesets[i]);
+    }
+    if (!startgids.empty()) {
+        m_gid_end = startgids.back() + tilesets.back()->total_tile_count();
+    }
+    std::sort(m_gid_map.begin(), m_gid_map.end(), order_by_gids);
+}
+
+std::pair<int, ConstTileSetPtr> GidTidTranslator::gid_to_tid(int gid) const {
+    if (gid == 0) {
+        return std::make_pair(0, nullptr);
+    }
+    if (gid < 1 || gid >= m_gid_end) {
+        throw InvArg{"Given gid is either the empty tile or not contained in "
+                     "this map; translatable gids: [1 " + std::to_string(m_gid_end)
+                     + ")."};
+    }
+    GidAndTileSetPtr samp;
+    samp.starting_id = gid;
+    auto itr = std::upper_bound(m_gid_map.begin(), m_gid_map.end(), samp, order_by_gids);
+    if (itr == m_gid_map.begin()) {
+        throw RtError{"Library error: GidTidTranslator said that it owned a "
+                      "gid, but does not have a tileset for it."};
+    }
+    --itr;
+    assert(gid >= itr->starting_id);
+    return std::make_pair(gid - itr->starting_id, itr->tileset);
+}
+
+void GidTidTranslator::swap(GidTidTranslator & rhs) {
+    m_gid_map.swap(rhs.m_gid_map);
+    std::swap(m_gid_end, rhs.m_gid_end);
+}
+
+/* private static */ bool GidTidTranslator::order_by_gids
+    (const GidAndTileSetPtr & lhs, const GidAndTileSetPtr & rhs)
+{ return lhs.starting_id < rhs.starting_id; }
